pyproject.toml
[project]
name = "fastapi-prod-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "black>=25.11.0",
    "fastapi[standard]>=0.122.0",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "ruff>=0.14.6",
    "sqlalchemy[asyncio]>=2.0.44",
]


[tool.ruff]

line-length = 88
target-version = "py311"


[tool.ruff.lint]

# Enable pyflakes (F), pycodestyle (E), isort (I)
select = ["E", "F", "I"]
ignore = ["F401", "E501"]  # Ignore unused imports everywhere



[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true



[tool.black]
max-line-length = 88 # Or your desired length
